// product-management-web/src/types/product.ts
export type Product = {
  id?: number;
  sku: string;
  name: string;
  description: string;
  category: string;
  price: number;
  stock: number;
  active: boolean;
};

export type ProductsPage = {
  content: Product[];
  totalElements: number;
  totalPages: number;
  size: number;
  number: number;
};